refactor: логика закрытия модального окна перенесена в компонент Modal
