+++ src/components/modal/modal.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import { CloseIcon } from '@ya.praktikum/react-developer-burger-ui-components';
@@ -7,6 +8,17 @@ import ModalOverlay from '../modal-overlay/modal-overlay';
 const modalRoot = document.getElementById('react-modals');
 
 function Modal({ children, title, closeAllPopups }) {
+  const handleEscPress = (e) => {
+    e.key === 'Escape' && closeAllPopups();
+  }
+
+  React.useEffect(() => {
+    document.addEventListener('keydown', handleEscPress);
+    return () => {
+      document.removeEventListener('keydown', handleEscPress)
+    }
+  });
+
   return ReactDOM.createPortal(
     <ModalOverlay closeAllPopups={closeAllPopups}>
       <div className={`${styles.modal} pl-10 pt-10 pr-10`}>
